Added an admin panel for management tasks (#1028)
* Fixed bootstrap import issue

* Added admin panel and user invite emails

* Remove error messages and content upon successfull invite

* Updated headers and padding

* Removed forceful alert styling

* Fixed alignment

* Restored headers

* Improved main card compatibility

* Added the ability to retrieve all users

* Fixed too wide test username

* Show is teacher

* Consistent style for card borders in grey

* Added is teacher to account creation

* Fixed merge leftovers

* Reduced header height

* Removed condensed font

* Added the ability to accept invites

* Prevent double request or other logged in user upon setting password

* Added the ability to remove users

* Replaced old theme blue color hardcodes

* Added the ability to make users a teacher

* Added the ability to import users from spreadsheet

* Fixed incorrect remove method

* Removed edit instance

* Fixed requested changes

* Added update instance to admin panel

* Fixed tests

* Fixed requested changes

* Improved toast

* Fixed reset button

* Added header height variable

* Fixed tests

* Fixed test

* Added course overview

* Updated button colors to new names

* Added user invite tests

* Added test for course get all

* Added docstring to invite users

* Improved feedback tests

* Added sender test

* Added search fields for course and user

* Fixed object toast issue

* Fixed whitespace issues

* More requested changes

* Even more

* Split test

* Moved header height var

* Use factory for invite tests

* Removed manual undo for bulk create

* Added data class for email sender names

* Also use email setting in test

* Added explicit string false param

* Added preferences initialisation to user bulk create

* Truncate too wide table headers

+++ src/django/test/test_user.py
@@ -1,3 +1,4 @@
+import re
 import test.factory as factory
 import test.factory.user as user_factory
 from test.utils import api
@@ -6,8 +7,11 @@ from test.utils.performance import QueryContext
 
 import pytest
 from django.conf import settings
+from django.contrib.auth.tokens import PasswordResetTokenGenerator
+from django.core import mail
 from django.core.exceptions import ValidationError
 from django.test import TestCase
+from django.test.utils import override_settings
 from rest_framework.settings import api_settings
 
 import VLE.factory as creation_factory
@@ -270,11 +274,9 @@ class UserAPITest(TestCase):
         # Test to delete user as other user
         api.delete(self, 'users', params={'pk': user2.pk}, user=user, status=403)
 
-        # Test to delete own user
-        api.delete(self, 'users', params={'pk': user.pk}, user=user)
-        api.get(self, 'users', params={'pk': user.pk}, user=admin, status=404)
-        api.delete(self, 'users', params={'pk': 0}, user=user2)
-        api.get(self, 'users', params={'pk': user2.pk}, user=admin, status=404)
+        # Test to delete own user (not possible, should still exist)
+        api.delete(self, 'users', params={'pk': user.pk}, user=user, status=403)
+        api.get(self, 'users', params={'pk': user.pk}, user=admin)
 
         # Test to delete user as admin
         api.delete(self, 'users', params={'pk': user3.pk}, user=admin)
@@ -498,3 +500,118 @@ class UserAPITest(TestCase):
                 values, prefetched = prefetched_objects(user, 'participation_set')
                 participation = values[0]  # Test is setup so there no need to check if prefetched with alternative qry
                 assert list(participation.groups.all()) == [group], 'All groups are prefetched for all users'
+
+    @override_settings(EMAIL_BACKEND='anymail.backends.test.EmailBackend', CELERY_TASK_ALWAYS_EAGER=True)
+    def test_invite_invalid_users(self):
+        admin = factory.Admin()
+        pre_invite_user_count = User.objects.count()
+        users = [{
+            'full_name': user.full_name,
+            'username': user.username,
+            'email': user.email,
+        } for user in [factory.Student.build() for _ in range(7)]]
+
+        users[0]['full_name'] = ''
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert 'Please specify a full name for all users. No invites sent.' in resp['description']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+        users[0]['username'] = ''
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert 'Please specify a username for all users. No invites sent.' in resp['description']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+        users[0]['email'] = ''
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert 'Please specify an email for all users. No invites sent.' in resp['description']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+        users[0]['username'] = users[1]['username']
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert users[0]['username'] in resp['description']['duplicate_usernames']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+        users[0]['email'] = users[1]['email']
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert users[0]['email'] in resp['description']['duplicate_emails']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+        # Username and email should be unique and case-insensitive.
+        # All-uppercase versions of existing values should thus not be accepted.
+        users[0]['username'] = User.objects.all().first().username.upper()
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert users[0]['username'].lower() in resp['description']['existing_usernames']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+        users[0]['email'] = User.objects.all().first().email.upper()
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=400)
+        assert users[0]['email'].lower() in resp['description']['existing_emails']
+        del users[0]
+        assert User.objects.count() == pre_invite_user_count, 'No accounts should be created if some user is invalid'
+
+    @override_settings(EMAIL_BACKEND='anymail.backends.test.EmailBackend', CELERY_TASK_ALWAYS_EAGER=True)
+    def test_invite_valid_users(self):
+        admin = factory.Admin()
+        pre_invite_user_count = User.objects.count()
+        pre_invite_teacher_count = User.objects.filter(is_teacher=True).count()
+        users = [{
+            'full_name': user.full_name,
+            'username': user.username,
+            'email': user.email,
+        } for user in [factory.Student.build() for _ in range(3)]]
+
+        # Add excess whitespace to some user details.
+        users[-1]['full_name'] = users[-1]['full_name'] + '   '
+        users[-1]['username'] = users[-1]['username'] + '   '
+        users[-1]['email'] = users[-1]['email'] + '   '
+
+        resp = api.post(self, 'users/invite_users', params={'users': users}, user=admin)
+        assert 'Successfully invited 3 users.' in resp['description']
+        assert User.objects.count() == pre_invite_user_count + 3
+        assert User.objects.filter(is_teacher=True).count() == pre_invite_teacher_count
+        assert not User.objects.last().full_name.endswith(' '), 'Whitespace should be removed from full names'
+        assert not User.objects.last().username.endswith(' '), 'Whitespace should be removed from usernames'
+        assert not User.objects.last().email.endswith(' '), 'Whitespace should be removed from emails'
+        assert all(Preferences.objects.filter(user__username=users[i]['username'].strip()).exists()
+                   for i in range(len(users))), 'All users should have preferences initialised after invitation'
+
+        assert len(mail.outbox) == 3, 'Invite email should be sent to all invited users'
+        assert all(mail.outbox[i].to == [users[i]['email'].strip()] for i in range(len(users))), \
+            'Email should be sent to the mail adress of each invited user'
+        assert all(users[i]['full_name'].strip() in mail.outbox[i].body for i in range(len(users))), \
+            'Invited user should be greeted by their full name'
+        assert all(f'{settings.BASELINK}/SetPassword/{users[i]["username"].strip()}/'
+                   in mail.outbox[i].alternatives[0][0] for i in range(len(users))), \
+            'Recovery token link should be in email'
+
+        # The recovery link should contain an indication that this is a new user, which is not part of the token
+        # itself.
+        for i in range(len(users)):
+            token = re.search(r'SetPassword\/(.*)\/([^"]*)\?new_user=true',
+                              mail.outbox[i].alternatives[0][0]).group(0).split('/')[-1][:-len('?new_user=true')]
+            assert PasswordResetTokenGenerator().check_token(User.objects.get(username=users[i]['username'].strip()),
+                                                             token), \
+                'Token for each user should be valid and new user should be indicated in URL'
+
+        resp = api.post(self, 'users/invite_users', params={'users': [{
+            'full_name': 'Invited Teacher Name',
+            'username': 'invitedteacherusername',
+            'email': 'invitedteacher@ejournal.app',
+            'is_teacher': True,
+        }]}, user=admin)
+        assert 'Successfully invited 1 users.' in resp['description']
+        assert User.objects.count() == pre_invite_user_count + 4
+        assert User.objects.filter(is_teacher=True).count() == pre_invite_teacher_count + 1
+
+        assert User.objects.filter(is_active=False).count() == 4, 'Invited user accounts are not yet active'
+
+        # Non-superusers may never invite new users.
+        admin.is_superuser = False
+        admin.save()
+        api.post(self, 'users/invite_users', params={'users': users}, user=admin, status=403)
